% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{choropleth}
\usage{
choropleth(fig, mapdata, group = NULL, pal = NULL,
  item_name = "users", decimals = NULL, suffix = "", na = NA)
}
\arguments{
\item{fig}{a leaflet object to which the choropleth polygons will be added}

\item{mapdata}{geojson (sf) object}

\item{suffix}{}
}
\description{
Choropleth map based on geojson data
}
\examples{
#
}
