% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{clean_kgraph}
\alias{clean_kgraph}
\title{Clean Knowledge Graph}
\usage{
clean_kgraph(kgraph)
}
\arguments{
\item{kgraph}{a dataframe or matrix of two columns.}
}
\value{
the cleaned version of  of the input, if no loops are found. If there are loops, will message and return NULL
}
\description{
Removes spurious and duplicated rows and checks for the absence of loops
}
\examples{
#
}
