% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{Cohen.d}
\alias{Cohen.d}
\title{Cohen's d of two samples}
\usage{
Cohen.d(x1 = NULL, x2 = NULL, m, sdev, n)
}
\arguments{
\item{x1}{first data sample, NAs will be ignored}

\item{x2}{second data sample, NAs will be ignored}

\item{m}{a vector of two values: means of the data samples (used only if \code{x1} or \code{x2} is \code{NULL})}

\item{sdev}{a vector of two values: standard deviations of the data samples (used only if \code{x1} or \code{x2} is \code{NULL})}

\item{n}{a vector of two values: sizes of the data samples (used only if \code{x1} or \code{x2} is \code{NULL})}
}
\value{
Cohen's d
}
\description{
Cohen's d of two samples
}
\examples{
#
}
