% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{col_add_a}
\alias{col_add_a}
\title{col_add_a}
\usage{
col_add_a(col, a = 0.5)
}
\arguments{
\item{col}{color codes}

\item{a}{transparency value on 0-to-1 scale}
}
\description{
Wrapper for adding transparency to a color
}
