% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{colorRampI}
\alias{colorRampI}
\title{colorRampI}
\usage{
colorRampI(x, na = "white", quantiles = TRUE, colors = c("white",
  "darkred"), ...)
}
\arguments{
\item{x}{a numerical vector to be mapped on colors}

\item{na}{if numeric - the value to which NAs in x are equivalent. If character - color for NA.}

\item{quantiles}{logical, whether or not map in quantile space, or directly.}

\item{colors}{- colors to interpolate, to be passed to colorRamp()}

\item{...}{other arguments to be passed to colorRamp()}
}
\value{
vector of color codes
}
\description{
A wrapper for colorRamp with a few modifications. Returns a vector of color codes
}
\examples{
#
}
