% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron.R
\name{cron}
\alias{cron}
\title{cron}
\usage{
cron(runfile, frequency = "1 day", start_time = Sys.time(),
  end_time = Inf, dt = 5, verbose = FALSE)
}
\arguments{
\item{runfile}{path to the R file to run}

\item{frequency}{a string consisting of an integer, a space, and then one of "seconds", "minutes", "hours", "days", "weeks", "months", "years". Plural form is optional and "sec" and "min" are understood.}

\item{start_time}{datetime object}

\item{end_time}{datetime object or Inf}

\item{dt}{pulse period in seconds}

\item{verbose}{logical}
}
\description{
A makeshift "cron job" functionality
}
\examples{
##
}
