% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_density.R
\name{gde}
\alias{gde}
\title{Gaussian-kernel density estimation}
\usage{
gde(x, xgrid = NULL, b = 1)
}
\arguments{
\item{x}{a matrix or data-frame of data points. Number of columns will be interpreted as dimensionality}

\item{xgrid}{a matrix or data-frame of point in which density will be computed, the same number of columns as in \code{x}. If NULL, will assume a uniform grid in the data range, 30 points in each dimension.}

\item{b}{a vector of bandwidths (standard deviations of the Gaussian kernel). If the length is not equal to the number of columns in \code{x}, will use the first element of \code{b} in all dimensions.}
}
\value{
a data-frame that is \code{xgrid} with the additional "density" column
}
\description{
Estimates density of data in arbitrary number of dimensions
}
