% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{getmode}
\alias{getmode}
\title{getmode}
\usage{
getmode(x, na.rm = FALSE)
}
\arguments{
\item{x}{- vector.}

\item{na.rm}{logical, whether to ignore NAs in x.}
}
\description{
Calculate the mode of a vector
}
