% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{gridclassifier}
\alias{gridclassifier}
\title{Grid-search through parameters of a classifier model}
\usage{
gridclassifier(data, model, formula, params, k = 1, repetitions = 1,
  verbose = FALSE)
}
\arguments{
\item{data}{the data frame of data, containing the predictor columns and the outcome column}

\item{model}{the function that executes the model, such as randomForest::randomForest or e1071::svm. The data frame argument of this function should be named "data", and the formula should be named "formula".}

\item{formula}{the formula of the classifier, to be passed to the model}

\item{params}{a list, each element is named as a parameter to be passed to the model and contains a vector of values to try.}

\item{k}{a positive integer, the number of folds in the k-fold validation}

\item{repetitions}{a positive integer, the number of repetitions in the k-fold validation}

\item{verbose}{a logical, whether or not to print out progress}
}
\value{
a list with components, all of them indexed in the same order: \item{class.err}{matrix of classification errors per class (each class is a column)}
\item{total.err}{a vector of total classification errors}
\item{confs}{a list of confusion matrices}
\item{gridparam}{a data frame of the grid of the parameters. It provides the connection between the index (common for all list components) and the parameter values.}
}
\description{
Given a vector of values of parameters, return classification errors and confusion matrices of a classifier model for each combination.
}
\details{
All parameters of the classifier model, out of which the grid will be formed, are assumed to be passed to the classifier as single values (not vectors, matrices, etc). If you want to grid-search parameters that your model takes as, say, a vector, or if your model takes the formula and the data as arguments with different names, use a wrapper function for your model.
}
\examples{
#Create fake data set and formula of prediction
data=data.frame(x=sample(c(0,1),100,replace=TRUE),p1=runif(100),p2=runif(100))
data$x=factor(data$x)
formula=as.formula('x~p1+p2')

#Create list of parameter values to try in the classifier
params=list(cw=c(0.7, 0.8, 0.9), cost=c(100, 500, 1000), gamma=c(1e-6, 0.01, 0.05, 1, 10))


#The SVM classifier e1071::svm takes class weights as a vector argument, so create a wrapper function
mysvm=function(formula, data, cw, ...){
   return(e1071::svm(formula=formula, data=data, class.weights=c('1'=cw, '0'=1-cw)))
}

#Find classification errors for each combination of parameters
ans=gridclassifier(data=data, model=mysvm, formula=formula, params=params)
}
