% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{isnothing}
\alias{isnothing}
\title{isnothing}
\usage{
isnothing(x, nothing = list(0, ""))
}
\arguments{
\item{x}{If length is greater than 1, will return TRUE if at least one element is nothing.}

\item{nothing}{list of values that count as nothing (used in addition to NA and the condition of positive length)}
}
\value{
TRUE or FALSE
}
\description{
Shorthand for checking if an object is nothing, e.g. of length 0 (including NULL), or NA, or '' or 0.
}
\examples{
#
}
