% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{kfoldind}
\alias{kfoldind}
\title{Sampling for k-fold validation}
\usage{
kfoldind(ndata, k = 5, repetitions = 3)
}
\arguments{
\item{ndata}{the number of rows in the data. If a vector is given, will use the length. If a matrix or a dataframe is given, will use number of rows.}

\item{k}{number of folds, positive integer.}

\item{repetitions}{number of repetitions of the k-fold procedure, positive integer.}
}
\value{
a list of \code{k*repetitions} components, each of which is a list of two vectors: indices of the training subset and of the validation subset.
}
\description{
Sampling for k-fold validation
}
\details{
Setting k=1 is understood as no folding: returned training and validation subsets will be both the full set.
}
\examples{
#
}
