% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{kfoldval}
\alias{kfoldval}
\title{Sampling for k-fold validation}
\usage{
kfoldval(ndata, k = 5, repetitions = 3)
}
\arguments{
\item{ndata}{the number of rows in the data}

\item{k}{number of folds}

\item{repetitions}{number of repetitions of the k-fold procedure}
}
\value{
a list of \code{k*repetitions} components, ieach of which is a vector of randomly sampled indices of length close to \code{ndata/k}.
}
\description{
Sampling for k-fold validation
}
