% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{kmeans1}
\alias{kmeans1}
\title{kmeans for 1D data with ordered clusters}
\usage{
kmeans1(x, centers, ...)
}
\arguments{
\item{x}{vector of 1D data}

\item{centers}{number of clusters}

\item{...}{other arguments passed to \code{kmeans}}
}
\value{
a list returned by \code{kmeans} with clusters arranged in the order of their means in the following list elements: \code{cluster, withinss, size, centers}.
Additional elements are also addded to the list: \code{frac} is the cluster size as a fraction of the total, \code{cluster_sd} is the standard deviations of clusters and \code{centers_se} is the standard error of the cluster centers.
}
\description{
Uses \code{kmeans} and then renames the clusters so that they are arranged in the increasing order of means
}
