% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2county.R
\name{latlong2county}
\alias{latlong2county}
\title{latlong2county}
\usage{
latlong2county(dat)
}
\arguments{
\item{dat}{tibble with  columns 'longitude' and 'latitude'}
}
\value{
dat with added columns 'fips','countyname', 'county_state' and 'county' (as polyname for package maps)
}
\description{
Map coordinates onto US counties
}
\examples{
#
}
