% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2county.R
\name{latlong2county_}
\alias{latlong2county_}
\title{latlong2county_}
\usage{
latlong2county_(df)
}
\arguments{
\item{df}{tibble}
}
\value{
tibble
}
\description{
Internal function, mapping coordinates onto US counties
}
\examples{
#
}
