% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{like_numeric}
\alias{like_numeric}
\title{like_numeric}
\usage{
like_numeric(x)
}
\arguments{
\item{x}{- vector}
}
\value{
logical vector of same length as x
}
\description{
Convenience function to determine if elements of a vector are convertible to numeric
}
\examples{
# like_numeric(c('1','a')) produces (TRUE, FALSE)
}
