% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyridges.R
\name{plotlyridges}
\alias{plotlyridges}
\title{Mimicks \code{ggridges} but using \code{plotly} so that the output is interactive}
\usage{
plotlyridges(data, vardens, varcat, linecolor = "darkblue",
  fillcolor = "steelblue", fillopacity = 0.6, linewidth = 0.5,
  scale = 0.9, logspaced = FALSE, cut.from = 0, cut.to = 3,
  n = 512, bw = NULL, bw.separate = FALSE,
  height.norm = "integral", round.digits = 2, x.min = 0,
  height = NULL, width = NULL)
}
\arguments{
\item{data}{data-frame of data to plot}

\item{vardens}{name of the column in \code{data} to use on the x-axis ('density variable')}

\item{varcat}{name of the column in \code{data} to use for splitting plots ('category variable')}

\item{linecolor}{line color}

\item{fillcolor}{fill color}

\item{fillopacity}{fill opacity}

\item{linewidth}{line width}

\item{scale}{vertical scale of plots, compared to the spacing between plots.}

\item{logspaced}{boolean, whether to use log-spaced points in calculating density}

\item{cut.from}{how much to cut into the region to the left of the smallest data values, in units of the bandwidth.}

\item{cut.to}{how much to cut into the region to the right of the greatest data values, in units of the bandwidth.}

\item{n}{number of points used in calculation of density}

\item{bw}{bandwidth to use in density calculation. if NULL, uses the default of \code{density}.}

\item{bw.separate}{if TRUE, will use separately estimated bandwidth in each plot, overriding \code{bw}}

\item{height.norm}{vertical normalization of plots. If 'integral', will normalize to unit area under the curve. If '1' will normalize to unit maximum height}

\item{round.digits}{number of rounding digits used in hover labels}

\item{x.min}{lower end of the x-axis range}

\item{height}{passed as \code{height} to the plotly object}

\item{width}{passed as \code{width} to the plotly object}
}
\value{
a plotly object
}
\description{
Mimicks \code{ggridges} but using \code{plotly} so that the output is interactive
}
\examples{
#
}
