% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{pN}
\alias{pN}
\title{pN}
\usage{
pN(x, suff = "", r = NULL)
}
\arguments{
\item{x}{vector of number}

\item{suff}{suffix to add}

\item{r}{number of digits after the decimal point. If NULL - no rounding.}
}
\value{
vector of strings
}
\description{
Pretty number. Rounding if r is not NULL. If abs(x) is less than 1.5, rounding changes keeping 2 significant digits.
}
