% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{pval}
\alias{pval}
\title{p-value for a 2-sample Welch's t-test.}
\usage{
pval(x1 = NULL, x2 = NULL, two.tailed = TRUE, m, sdev, n)
}
\arguments{
\item{x1}{first data sample, NAs will be ignored}

\item{x2}{second data sample, NAs will be ignored}

\item{two.tailed}{boolean to indicate whether the two-tailed or single-tailed p-value should be calculated}

\item{m}{a vector of two values: means of the data samples (used only if \code{x1} or \code{x2} is \code{NULL})}

\item{sdev}{a vector of two values: standard deviations of the data samples (used only if \code{x1} or \code{x2} is \code{NULL})}

\item{n}{a vector of two values: sizes of the data samples (used only if \code{x1} or \code{x2} is \code{NULL})}
}
\value{
the Welch's t-test p-value
}
\description{
Can be calculated from the provided samples. Alternatively, if either one of the samples is NULL, can be calculated from the provided means, standard deviations and sample sizes
}
\examples{
#
}
