% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid_plot}
\alias{pyramid_plot}
\title{pyramid_plot}
\usage{
pyramid_plot(dat, yticks = NULL, xtitle = "", ytitle = "",
  yname = NULL, palette = c("tomato", "steelblue"),
  plot_bgcolor = "#ECF0F5", paper_bgcolor = "#FCFCFC")
}
\arguments{
\item{dat}{tibble with columns "age" and "gender". "gender" should be a factor with two levels (e.g. "Female" and "Male"). If not a factor, will be converted into a factor. Age can be numeric or character.}

\item{yticks}{values to show on the yaxis}

\item{xtitle}{x-axis title}

\item{ytitle}{y-axis title}

\item{yname}{name for the variable on y-axis, to show in the hoverplate.}

\item{palette}{vector of two colors, for Female and Male (in that order).}

\item{plot_bgcolor}{plot background color}

\item{paper_bgcolor}{plot paper color}
}
\value{
plotly object
}
\description{
Create a population pyramid plot. Traditionally, x-axis is gender and y-axis is age, so even though in this plot they both can have other names and values, we use "age" and "gender" for column names.
}
\examples{
#
}
