% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_plot}
\alias{segment_plot}
\title{segment_plot}
\usage{
segment_plot(dat, n = 20, unit = "", xtitle = "", line_width = 3,
  arrow_size = 7, plot_bgcolor = "#ECF0F5",
  paper_bgcolor = "#FCFCFC")
}
\arguments{
\item{dat}{tibble with columns label (values for the y axis), x and xend (these two have to be numeric)}

\item{n}{integer. If positive, # of labels with the largest change to plot. If negative, # of labels with the smallest change to plot}

\item{unit}{char, units to be shown in the hoverplates}

\item{xtitle}{x-axis title}

\item{line_width}{line width}

\item{arrow_size}{arrow size}

\item{plot_bgcolor}{plot background color}

\item{paper_bgcolor}{plot paper color}
}
\value{
plotly object
}
\description{
Segment plot, showing an arrow from x to xend for each value of y
}
\examples{
#
}
