% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{sortnum}
\alias{sortnum}
\title{Sorts string vector by numeric characters}
\usage{
sortnum(x, ...)
}
\arguments{
\item{x}{vector of strings}

\item{...}{other arguments passed to the function \code{sort}}
}
\value{
sorted vector of strings
}
\description{
Numeric characters in each vector element are concatenated and treated as integers, and the sorting is done by these intergers. E.g. \code{c('a10','a2','a1')} would be sorted as \code{c('a1','a2','a10')}.
}
