% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst_plot}
\alias{sunburst_plot}
\title{sunburst_plot}
\usage{
sunburst_plot(dat, item_name, domain = list(x = c(0, 1), y = c(0, 1)))
}
\arguments{
\item{dat}{tibble, as prepared by sunburstdat()}

\item{item_name}{string}

\item{domain}{as in plotly}
}
\value{
plotly object
}
\description{
Create sunburst plot
}
\examples{
#
}
