% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburstdat}
\alias{sunburstdat}
\title{sunburstdat}
\usage{
sunburstdat(dat, pal = NULL)
}
\arguments{
\item{dat}{tibble}

\item{pal}{palette - a vector of colors. If NULL, will use Set1 from RColorBrewer}
}
\value{
a tibble that can be passed to plotly(type='sunburst')
}
\description{
Prepares data for plotly geographic sunburst plot.
}
\examples{
#
}
