% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline_intervals}
\alias{timeline_intervals}
\title{Timeline_intervals}
\usage{
timeline_intervals(dat, startcol = "start_time", endcol = "end_time",
  trange = NULL, step = 15, units = "mins", time_zone = "UTC")
}
\arguments{
\item{dat}{tibble}

\item{startcol}{name of the column with start-times (quoted)}

\item{endcol}{name of the column with end-times (quoted)}

\item{trange}{range of times. If NULL, will be inferred from data.}

\item{step}{numeric, timestep to use in binning}

\item{units}{character, units for step, as in difftime(), e.g. 'mins' or 'days'}
}
\value{
a tibble with columns t (bin-end times), n - count of ongoing events in the bin
}
\description{
An extension of timeline(summarize='n', crop=FALSE) for intervals rather than single events.
}
