% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline_plot}
\alias{timeline_plot}
\title{timeline_plot}
\usage{
timeline_plot(dat, smooth_bw = 0, smooth_kernel = "box",
  ytitle = "count", plot_bgcolor = "#ECF0F5",
  paper_bgcolor = "#FCFCFC", linecolor = "darkgreen",
  fillcolor = "steelblue", hovermode = "x", ticksuffix = "",
  ytitle_fontsize = 14, sliderthickness = 0.05, yrange = NULL,
  linenames = NULL)
}
\arguments{
\item{dat}{tibble with columns t, n and n_base (optional)}

\item{smooth_bw}{bandwidth for kernel-smoothing the curve, in units of the time step in the data. If 0 - no smoothing.}

\item{smooth_kernel}{kernel used in smoothing the curve: 'box' or 'normal'.}

\item{ytitle}{title for the y-axis}

\item{plot_bgcolor}{plot background color}

\item{paper_bgcolor}{plot paper color}

\item{linecolor}{line color}

\item{fillcolor}{fill color}

\item{hovermode}{passed to plotly}

\item{ticksuffix}{tick suffix}

\item{ytitle_fontsize}{fontsize for the y-axis title}

\item{sliderthickness}{thickness (relative height) of the range slider}

\item{yrange}{range for the y-axis. If NULL, will be determined from the data}

\item{linenames}{name for the line(s), to show in the hoverplates. If n_base is present in the data, should be length 2.}
}
\value{
plotly object
}
\description{
Create a timeline plot with one (n(t)) or two (if n_base is present in the data) lines.
}
