% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron.R
\name{tnext_}
\alias{tnext_}
\title{tnext_}
\usage{
tnext_(t, tnow = Sys.time(), frequency = "1 day")
}
\arguments{
\item{t}{datetime, start of sequence}

\item{frequency}{a string consisting of an integer, a space, and then one of "seconds", "minutes", "hours", "days", "weeks", "months", "years". Plural form is optional and "sec" and "min" are understood.}
}
\value{

}
\description{
Given a time sequence from the time t with a frequency, return the first sequence element that is >= tnow
}
