% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{truncateString}
\alias{truncateString}
\title{truncateString}
\usage{
truncateString(x, n = 15)
}
\arguments{
\item{x}{vector of strings}

\item{n}{max number of characters}
}
\value{
vector of strings
}
\description{
A cosmetic function for truncating a string to n characters, adding "..." if needed.
}
\examples{
#
}
