% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{visdat}
\alias{visdat}
\title{visdat}
\usage{
visdat(dat, level = "us_county", us_counties = NULL,
  country_name_alphacode = NULL)
}
\arguments{
\item{dat}{tibble.}

\item{level}{either 'us_county' (contiguous us map on the level of counties), or 'world' (on the level of countries).}
}
\value{
a list object that can be passed to leaflet()
}
\description{
Prepares data for a choropleth leaflet map.
}
