% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{visdat_}
\alias{visdat_}
\title{visdat_}
\usage{
visdat_(dat, level = "us_county", vars = "n")
}
\arguments{
\item{dat}{tibble.}

\item{level}{either 'us_county' (contiguous us map on the level of counties), or 'world' (on the level of countries).}
}
\value{
a list object that can be passed to leaflet()
}
\description{
Internal function. Prepares data for a choropleth leaflet map.
}
\examples{
#
}
