% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{choropleth}
\usage{
choropleth(fig, mapdata, group = NULL, pal = NULL, na = NA)
}
\arguments{
\item{fig}{a leaflet object to which the choropleth polygons will be added.}

\item{mapdata}{geojson (sf or sp) object. Should contain columns n, label, color.}

\item{na}{- value with which to replace NAs in n.}
}
\description{
Choropleth map based on geojson data
}
\examples{
#
}
