% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron.R
\name{cron}
\alias{cron}
\title{cron}
\usage{
cron(runfile, frequency = "1 day", start_time = Sys.time(),
  only_on = NULL, end_time = Inf, dt = 5, verbose = FALSE)
}
\arguments{
\item{runfile}{path to the R file to run}

\item{frequency}{a string consisting of an integer, a space, and then one of "seconds", "minutes", "hours", "days", "weeks", "months", "years". Plural form is optional and "sec" and "min" are understood.}

\item{start_time}{datetime object}

\item{only_on}{a vector of weekdays (full names capitalized, e.g. "Sunday"), on which to run. An object of 0 length (e.g. NULL) is equivalent to no restriction.}

\item{end_time}{datetime object or Inf}

\item{dt}{pulse period in seconds. Note that dt should be less than frequency.}

\item{verbose}{logical}
}
\description{
A makeshift "cron job" functionality
}
\examples{
##
}
