% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{gridclassifier}
\alias{gridclassifier}
\title{Grid-search through parameters of a classifier model}
\usage{
gridclassifier(data, model, formula, params, k = 1, repetitions = 1,
  verbose = FALSE, writeto = NULL)
}
\arguments{
\item{data}{the data frame of data, containing the predictor columns and the outcome column}

\item{model}{the function that executes the model, such as randomForest::randomForest or e1071::svm. The data frame argument of this function should be named "data", and the formula should be named "formula".}

\item{formula}{the formula of the classifier, to be passed to the model}

\item{params}{a list, each element is named as a parameter to be passed to the model and contains a vector of values to try.}

\item{k}{a positive integer, the number of folds in the k-fold validation}

\item{repetitions}{a positive integer, the number of repetitions in the k-fold validation}

\item{verbose}{a logical, whether or not to print out progress}

\item{writeto}{if not NULL, will rewrite results cumulatively after each combination of parameters is finished to the file with this path.}
}
\value{
a list with components, indexed in the same order: \item{class.err}{matrix of classification errors per class (each class is a column)}
\item{total.err}{a vector of total classification errors}
\item{confs}{a list of confusion matrices}
\item{gridparam}{a data frame of the grid of the parameters. It provides the connection between the index (common for the preceding list components) and the parameter values.}
\item{kfold}{a list containing the number of folds and repetitions in the kfold validation}
}
\description{
Given a vector of values of parameters, return classification errors and confusion matrices of a classifier model for each combination.
}
\details{
All parameters of the classifier model, out of which the grid will be formed, are assumed to be passed to the classifier as single values (not vectors, matrices, etc). If you want to grid-search parameters that your model takes as, say, a vector, or if your model takes the formula and the data as arguments with different names, use a wrapper function for your model.
}
