% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{map_world}
\alias{map_world}
\title{map_world}
\usage{
map_world(dat, return_df = FALSE, include_vars = c("n"))
}
\arguments{
\item{dat}{tibble, containing country (2-letter country codes, e.g. "US"), n (numeric variable to be plotted) and label (to be shown in hoverlabel)}

\item{return_df}{#Boolean, whether to return table of results instead of a plot}

\item{include_vars}{list of variables to include in the returned table. If return_df is FALSE, ignored.}

\item{item_name}{string, description of the plotted variable, to be shown in the hoverlabel}

\item{suffix}{string, units to be shown after the value of n in the hoverlabel}

\item{decimals}{# integer, decimal points in rounding}
}
\value{
leaflet object or tibble, depending on return_df
}
\description{
Choropleth map of the world on country level
}
